Added delete confirmation and edit/submit alerts and setlist functionality through the admin page which then displays on the setlist page

+++ library/views.py
@@ -1,7 +1,8 @@
 from django.http import HttpResponseRedirect
 from django.shortcuts import get_object_or_404, render
 from django.contrib.auth.decorators import login_required
-from django.urls import reverse
+from django.contrib import messages
+from django.urls import resolve, reverse
 
 from .forms import PieceForm
 from .models import Piece
@@ -15,12 +16,15 @@ def library_view(request):
         if piece_form.is_valid():
             piece_form.save()
             return HttpResponseRedirect(reverse("library"))
-        
+
     pieces = Piece.objects.all()
     piece_form = PieceForm()
-    return render(
-        request, "library/library.html", {"pieces": pieces, "piece_form": piece_form}
-    )
+    context = {
+        "pieces": pieces,
+        "piece_form": piece_form,
+        "current_url": resolve(request.path_info).url_name,
+    }
+    return render(request, "library/library.html", context)
 
 
 def piece_edit(request, pk):
@@ -29,17 +33,23 @@ def piece_edit(request, pk):
         piece_form = PieceForm(request.POST, instance=piece)
         if piece_form.is_valid():
             piece_form.save()
+            messages.success(request, "Update Successful")
             return HttpResponseRedirect(reverse("library"))
 
     piece = get_object_or_404(Piece, pk=pk)
     piece_form = PieceForm(instance=piece)
     pieces = Piece.objects.all()
-    return render(
-        request, "library/library.html", {"pieces": pieces, "piece_form": piece_form}
-    )
+    context = {
+        "pieces": pieces,
+        "piece_form": piece_form,
+        "current_url": resolve(request.path_info).url_name,
+    }
+    return render(request, "library/library.html", context)
 
 
 def piece_delete(request, pk):
     piece = get_object_or_404(Piece, pk=pk)
-    piece.delete()
+    if request.method == "POST":
+        piece.delete()
+        messages.success(request, "Deletion Successful")
     return HttpResponseRedirect(reverse("library"))
